backend/requirements.txt
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.2
PyJWT 
bcrypt
tensorflow==2.11.0
transformers
torch
torchvision
torchaudio
matplotlib
boto3